+++ src/main/scala/com/crobox/clickhouse/ClickhouseException.scala
@@ -1,8 +1,10 @@
 package com.crobox.clickhouse
 
+import akka.http.scaladsl.model.StatusCode
+
 /**
  * @author Sjoerd Mulder
  * @since 31-03-17
  */
-class ClickhouseException(message: String, query: String, cause: Throwable = null)
+case class ClickhouseException(message: String, query: String, cause: Throwable = null, statusCode: StatusCode)
     extends RuntimeException(message + s", query $query", cause)
